Remove 5.1.0 from bad_versions
Remove the 5.1.0 release of vddk from the list of unsupported versions.
VMware has stopped shipping the vddk with patch releases other than ".0"
They ship some versions as 5.1.0 update N, and some versions as 5.1.x
where x is a non-zero number, but the .so library files will have the
"5.1.0" suffix in all cases.  The code checking the supported versions
can only be used going forward if we stop supporting an entire point release.

https://bugzilla.redhat.com/show_bug.cgi?id=1151465